Added support for PNG pictures for cursors.
A new attribute is added to the style sheet to support loading png files.
"imgType"  attribute supports  "cur" or "png".

Basically, if empty or "cur" in the field will cause the cursor data to be loade through JMECursor.  Otherwise it will define a new JMECursor and load the PNG as a texture and convert it to the right format and set the image within JMECursor.